+++ Color.lua
@@ -4,6 +4,8 @@ local Object = require(BASE .. "Object")
 --- @type loveshape.utils
 local utils = require(BASE .. "utils")
 
+local clamp = utils.clamp
+
 --- Represents an RGBA color. Each color component is given as a floating point
 --- value in the range from 0 to 1.
 ---
@@ -22,10 +24,10 @@ function Color.new(r, g, b, a)
     r, g, b, a = unpack(r)
   end
   return setmetatable({
-    utils.clamp(assert(r, "red required"), 0, 1),
-    utils.clamp(assert(g, "green required"), 0, 1),
-    utils.clamp(assert(b, "blue required"), 0, 1),
-    utils.clamp(a or 1, 0, 1)
+    clamp(assert(r, "red required"), 0, 1),
+    clamp(assert(g, "green required"), 0, 1),
+    clamp(assert(b, "blue required"), 0, 1),
+    clamp(a or 1, 0, 1)
   }, Color)
 end
 
@@ -40,10 +42,10 @@ function Color:set(r, g, b, a)
     r, g, b, a = unpack(r)
   end
 
-  self[1] = utils.clamp(assert(r, "red required"), 0, 1)
-  self[2] = utils.clamp(assert(g, "green required"), 0, 1)
-  self[3] = utils.clamp(assert(b, "blue required"), 0, 1)
-  self[4] = utils.clamp(a or 1, 0, 1)
+  self[1] = clamp(assert(r, "red required"), 0, 1)
+  self[2] = clamp(assert(g, "green required"), 0, 1)
+  self[3] = clamp(assert(b, "blue required"), 0, 1)
+  self[4] = clamp(a or 1, 0, 1)
 
   return self
 end
@@ -55,6 +57,20 @@ end
 --- @nodiscard
 function Color:unpack()
   return self[1], self[2], self[3], self[4]
+end
+
+--- Determine if the color is equal to another color.
+--- @param r number
+--- @param g number
+--- @param b number
+--- @param a number? (default: 1)
+--- @return boolean
+--- @overload fun(self: loveshape.Color, rgba: table): boolean
+function Color:equal(r, g, b, a)
+  if type(r) == "table" then
+    r, g, b, a = unpack(r)
+  end
+  return self[1] == r and self[2] == g and self[3] == b and self[4] == (a or 1)
 end
 
 return Color