+++ Ellipse.lua
@@ -4,14 +4,19 @@ local Shape = require(BASE .. "Shape")
 --- @type loveshape.utils
 local utils = require(BASE .. "utils")
 
+local TWO_PI = math.pi * 2
+
 --- @class loveshape.Ellipse: loveshape.Shape
 --- @field protected _radiusX number Horizontal radius
 --- @field protected _radiusY number Vertical radius
 local Ellipse = utils.class("loveshape.Ellipse", Shape)
 
+--- Create a new ellipse with the specified dimensions.
+---
+--- If `pointsCount` is not specified, a default value is computed automatically.
 --- @param radiusX number Horizontal radius
 --- @param radiusY number Vertical radius
---- @param pointsCount integer Number of points used to draw the ellipse
+--- @param pointsCount? integer Number of points used to draw the ellipse
 --- @return loveshape.Ellipse
 --- @nodiscard
 function Ellipse.new(radiusX, radiusY, pointsCount)
@@ -23,11 +28,15 @@ end
 
 --- @param radiusX number
 --- @param radiusY number
---- @param pointsCount integer
+--- @param pointsCount? integer
 --- @protected
 function Ellipse:_init(radiusX, radiusY, pointsCount)
   assert(type(radiusX) == "number" and radiusX > 0)
   assert(type(radiusY) == "number" and radiusY > 0)
+
+  if not pointsCount then
+    pointsCount = utils.computeEllipsePoints(radiusX, radiusY)
+  end
 
   Shape._init(self, pointsCount)
   self:setRadius(radiusX, radiusY)
@@ -74,7 +83,7 @@ function Ellipse:getPoint(index)
     error("Invalid point index: " .. tostring(index))
   end
 
-  local angle = (math.pi * 2) / pointsCount
+  local angle = TWO_PI / pointsCount
   local x, y = utils.vecFromAngle(angle * (index - 1))
 
   return x * self._radiusX, y * self._radiusY