+++ CHANGES.md
@@ -3,6 +3,21 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-12-02
+
+### Added
+
+- Small code optimizations
+- `Color:equal` to check if two colors are equal
+- `utils.computeEllipsePoints` to compute the optimal number of points to draw an ellipse or circle
+
+### Changed
+
+- Only update meshes in `Shape:setFillColor` and `Shape:setBorderColor` when a new color is set
+- Compute `pointsPerCorner` automatically if not specified in `RoundedRectangle.new`
+- Compute `pointsCount` automatically if not specified in `Circle.new` and `Ellipse.new`
+- Replace local loveunit module with lovecase for unit tests
+
 ## [1.1.0] - 2023-11-20
 
 ### Added